Design: 검색 페이지 CSS 수정
검색 페이지 CSS 수정
